Fix empty state for ImagePanel
As an optimization, the image panel always keeps the ImageCanvas mounted even when there is no current image to display. When seeking or changing the message order, there may not be an image loaded for display, yet the image panel would erroneously show the stale image alongside the empty state message.

This change fixes the styles so the empty state covers any stale image and avoid split screen with empty state and stale image.
